backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
google-cloud-secret-manager
 
## sqlalchemy: É o ORM (Object-Relational Mapper) que nos permitirá interagir com o banco de dados usando objetos Python em vez de SQL puro.
## psycopg2-binary: É o "driver" que permite que a SQLAlchemy se comunique especificamente com o PostgreSQL.
## alembic: É a ferramenta que usaremos para gerir as migrações do esquema do banco de dados (criar tabelas, alterar colunas, etc.).
## python-dotenv: Ajuda a carregar variáveis de ambiente de um ficheiro .env para o desenvolvimento local.
## google-cloud-secret-manager: Biblioteca da Google para aceder a segredos (como a nossa senha do banco) de forma segura.
